JINJA TEMPLATING, LANGUAGE & HTML TEMPLATES
Create my HTML pages to display user information.
exploiting my base template as it will be the "Theme" of my website, also touching on the topic of "Template Inheritance", the parts of this template will be over-ridden by the HTML DOCS that inherit from this base template (base.html) document with their respective specific content.