JINJA TEMPLATING, LANGUAGE & HTML TEMPLATES
How to actually pass values to templates?
What is good with "Jinja" is that we can pass values right to all of our templates, That will allow us to use our template.
e.g. inside our login.html template we can access the variable "text".
As you can see in my login.html i have passed the variable "text" so you see that whatever is passed with the text to this template will show up.